+++ launch/arm_controller.launch
@@ -5,6 +5,7 @@
         <node pkg="robot_arm_workcell_manager" type="robot_arm_controller" name="robot_arm_workcell_manager" respawn="true">
             <!-- Note: param cant use ${pkg}, thus use arg to solve this prob -->
             <param name="config_yaml_path" type="string" value="$(arg file)" />
+            <param name="group_name" type="string" value="manipulator" />
         </node>
     </group>
 </launch>
